path_operations_structure_fixed
