+++ endpoints/__init__.py
@@ -1,0 +1,3 @@
+from .inicial_router import inicial_router
+from .users_router import user_router
+from .tweets_router import tweet_router
