+++ main.py
@@ -1,11 +1,9 @@
 from fastapi import FastAPI
 
-from endpoints.inicial_routes import inicial_routes
-from endpoints.users_routes import users_routes
-from endpoints.tweets_routes import tweets_routes
+from endpoints import inicial_router, user_router, tweet_router
 
 app = FastAPI()
 
-app.include_router(inicial_routes)
-app.include_router(users_routes)
-app.include_router(tweets_routes)
+app.include_router(inicial_router)
+app.include_router(user_router)
+app.include_router(tweet_router)
