+++ models/__init__.py
@@ -1,10 +1,3 @@
+# Models to import from models/
 from .user import User, userLogin
 from .tweet import Tweet
-
-# Pydantic
-from pydantic import BaseModel, Field
-
-# Python Native Library
-from datetime import date, datetime
-from uuid import UUID
-from typing import Optional
